+++ frontend/packages/sheet/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./provider";
 export * from "./components";
+export * from "./helpers";
